I know this doesn't work but
Some of this code is right. But I think I just now realized that I need to make a clear separation of each dragging process, where each tile drag starts and ends, when I select different tiles, when I undo and move the selected tiles. I need a way to go back to each state. So create some kind of SelectionDragState or something.